Separate failure_detector to file
